refactor(modules): extract `bichat` from `core` into its own module
