+++ modules/bichat/domain/entities/dialogue/dialogue_events.go
@@ -2,8 +2,10 @@ package dialogue
 
 import (
 	"context"
+
 	"github.com/iota-uz/iota-sdk/modules/core/domain/aggregates/user"
 	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/session"
+
 	"github.com/iota-uz/iota-sdk/pkg/composables"
 )
