+++ modules/bichat/domain/entities/dialogue/dialogue_impl.go
@@ -1,8 +1,9 @@
 package dialogue
 
 import (
-	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/llm"
 	"time"
+
+	"github.com/iota-uz/iota-sdk/modules/bichat/domain/entities/llm"
 )
 
 func New(userID uint, label string) Dialogue {
