+++ modules/bichat/domain/entities/dialogue/dialogue_reply.go
