+++ modules/bichat/domain/entities/dialogue/dialogue_repository.go
@@ -1,8 +1,6 @@
 package dialogue
 
-import (
-	"context"
-)
+import "context"
 
 type FindParams struct {
 	Query  string
