+++ modules/bichat/domain/entities/dialogue/dialogue_start.go
