+++ modules/bichat/domain/entities/embedding/embedding.go
