+++ modules/bichat/domain/entities/llm/llm.go
