+++ modules/bichat/domain/entities/llm/message.go
