+++ modules/bichat/domain/entities/prompt/prompt.go
@@ -1,8 +1,6 @@
 package prompt
 
-import (
-	"time"
-)
+import "time"
 
 type Prompt struct {
 	ID          string
