+++ modules/bichat/domain/entities/prompt/prompt_repository.go
@@ -1,8 +1,6 @@
 package prompt
 
-import (
-	"context"
-)
+import "context"
 
 type Repository interface {
 	Count(ctx context.Context) (int64, error)
