+++ modules/bichat/infrastructure/llmproviders/openai_provider.go
@@ -2,7 +2,9 @@ package llmproviders
 
 import (
 	"context"
-	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/llm"
+
+	"github.com/iota-uz/iota-sdk/modules/bichat/domain/entities/llm"
+
 	"github.com/sashabaranov/go-openai"
 )
