+++ modules/bichat/module.go
@@ -1,0 +1,40 @@
+package bichat
+
+import (
+	"embed"
+
+	"github.com/iota-uz/iota-sdk/modules/bichat/infrastructure/persistence"
+	"github.com/iota-uz/iota-sdk/modules/bichat/presentation/controllers"
+	"github.com/iota-uz/iota-sdk/modules/bichat/services"
+	"github.com/iota-uz/iota-sdk/pkg/application"
+	"github.com/iota-uz/iota-sdk/pkg/spotlight"
+)
+
+//go:embed presentation/locales/*.json
+var LocaleFiles embed.FS
+
+func NewModule() application.Module {
+	return &Module{}
+}
+
+type Module struct {
+}
+
+func (m *Module) Register(app application.Application) error {
+	app.RegisterLocaleFiles(&LocaleFiles)
+	app.RegisterServices(
+		services.NewEmbeddingService(app),
+		services.NewDialogueService(persistence.NewDialogueRepository(), app),
+	)
+	app.RegisterControllers(
+		controllers.NewBiChatController(app),
+	)
+	app.Spotlight().Register(
+		spotlight.NewItem(nil, BiChatLink.Name, BiChatLink.Href),
+	)
+	return nil
+}
+
+func (m *Module) Name() string {
+	return "bichat"
+}
