+++ modules/bichat/presentation/locales/ru.json
@@ -1,0 +1,19 @@
+{
+  "NavigationLinks": {
+    "BiChat": "BI-Чат"
+  },
+  "BiChat": {
+    "Meta": {
+      "Index": {
+        "Title": "BI-Чат"
+      }
+    },
+    "Title": "BI-Чат",
+    "NewDialog": "Новый диалог",
+    "MessagePlaceholder": "Введите сообщение...",
+    "History": {
+      "Empty": "История пуста",
+      "HelpText": "Нажмите на кнопку \"Новый диалог\" чтобы начать новый чат"
+    }
+  }
+}
