+++ modules/bichat/services/chatfuncs/currency_convert.go
