+++ modules/bichat/services/embeddings_service.go
@@ -5,11 +5,13 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/embedding"
-	"github.com/iota-uz/iota-sdk/pkg/application"
-	"github.com/iota-uz/iota-sdk/pkg/llm/gpt-functions"
 	"log"
 	"net/http"
+
+	"github.com/iota-uz/iota-sdk/modules/bichat/domain/entities/embedding"
+	functions "github.com/iota-uz/iota-sdk/pkg/llm/gpt-functions"
+
+	"github.com/iota-uz/iota-sdk/pkg/application"
 )
 
 type EmbeddingService struct {
