+++ modules/bichat/services/prompt_service.go
@@ -2,7 +2,8 @@ package services
 
 import (
 	"context"
-	prompt2 "github.com/iota-uz/iota-sdk/modules/core/domain/entities/prompt"
+
+	prompt2 "github.com/iota-uz/iota-sdk/modules/bichat/domain/entities/prompt"
 	"github.com/iota-uz/iota-sdk/pkg/eventbus"
 )
