+++ modules/core/infrastructure/persistence/models/models.go
@@ -153,14 +153,6 @@ type UserRole struct {
 	CreatedAt time.Time
 }
 
-type Prompt struct {
-	ID          string
-	Title       string
-	Description string
-	Prompt      string
-	CreatedAt   time.Time
-}
-
 type EmployeeContact struct {
 	ID         uint
 	EmployeeID uint
@@ -298,15 +290,6 @@ type ActionLog struct {
 	UserAgent string
 	IP        string
 	CreatedAt time.Time
-}
-
-type Dialogue struct {
-	ID        uint
-	UserID    uint
-	Label     string
-	Messages  string
-	CreatedAt time.Time
-	UpdatedAt time.Time
 }
 
 type Session struct {
