+++ modules/core/presentation/locales/ru.json
@@ -146,7 +146,6 @@
     "Administration": "Администрирование",
     "Users": "Пользователи",
     "Roles": "Роли",
-    "BiChat": "BI-Чат",
     "KnowledgeBase": "База знаний",
     "Operations": "Операции",
     "Employees": "Сотрудники",
@@ -195,20 +194,6 @@
       "Title": "Панель управления"
     },
     "Welcome": "Добро пожаловать на панель управления"
-  },
-  "BiChat": {
-    "Meta": {
-      "Index": {
-        "Title": "BI-Чат"
-      }
-    },
-    "Title": "BI-Чат",
-    "NewDialog": "Новый диалог",
-    "MessagePlaceholder": "Введите сообщение...",
-    "History": {
-      "Empty": "История пуста",
-      "HelpText": "Нажмите на кнопку \"Новый диалог\" чтобы начать новый чат"
-    }
   },
   "Users": {
     "Meta": {
