+++ pkg/llm/gpt-functions/funcs.go
@@ -2,7 +2,8 @@ package functions
 
 import (
 	"encoding/json"
-	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/llm"
+
+	"github.com/iota-uz/iota-sdk/modules/bichat/domain/entities/llm"
 
 	"gorm.io/gorm"
 )
